refactor(session): move samples assembly and codec into a samples/ package
samples.py -> samples/merge.py and samples_codec.py -> samples/codec.py, so
the assembly and its wire codec sit as siblings under one package instead of
two prefixed top-level modules; import sites updated mechanically.

Also trim the codec module docstring to the black-box contract (what goes in,
what comes out of encode/decode); the wire-format details stay in the inline
comments next to the code they constrain.
